Extracted dbWrapper from session.controller
