+++ server/app/controllers/session.controller.js
@@ -7,47 +7,17 @@ const { filterUserAttributes } = require('app/helpers/usersHelper');
 module.exports = (db) => {
   const database = db.dynamoDb;
   const userTable = db.users;
+  const dbWrapper = db.dbWrapper;
 
   const signin = async (req, res) => {
     try {
-      const searchForUsername = {
-        TableName: userTable,
-        IndexName: 'usernameIndex',
-        KeyConditionExpression: '#username = :username',
-        ExpressionAttributeNames: {
-          '#username': 'username',
-        },
-        ExpressionAttributeValues: {
-          ':username': req.body.user,
-        },
-      };
-      const searchForUserEmail = {
-        TableName: userTable,
-        IndexName: 'emailIndex',
-        KeyConditionExpression: '#email = :email',
-        ExpressionAttributeNames: {
-          '#email': 'email',
-        },
-        ExpressionAttributeValues: {
-          ':email': req.body.user,
-        },
-      };
-      const dbRequests = [
-        database.query(searchForUsername).promise(),
-        database.query(searchForUserEmail).promise(),
-      ];
-      const results = await Promise.all(dbRequests);
-      const usernameResult = results[0];
-      const userEmailResult = results[1];
+      const user = await dbWrapper.searchForUser(req.body.user)
 
-      if (usernameResult.Count === 0 && userEmailResult.Count === 0) {
+      if (!user) {
         return res.status(404).send({
           message: 'User not found',
         });
       }
-      const user = usernameResult.Count > 0
-        ? usernameResult.Items[0]
-        : userEmailResult.Items[0];
 
       const isPasswordValid = await bcrypt.compare(
         req.body.password,
