+++ server/server.js
@@ -1,10 +1,11 @@
 const serverless = require('serverless-http');
 const AWS = require('aws-sdk');
+const dbWrapper = require('./app/db/dbWrapper')
 
 // setup db
 const { USERS_TABLE } = process.env;
 const dynamoDb = new AWS.DynamoDB.DocumentClient();
-const db = { dynamoDb, users: USERS_TABLE };
+const db = { dynamoDb, users: USERS_TABLE, dbWrapper: dbWrapper(dynamoDb, USERS_TABLE) };
 
 // create app
 const app = require('./app')(db);
